[refactor/#224] MyAtivities 폴더명 오타 및 import 경로 수정
