+++ src/pages/activity/[[...activityId]]/index.tsx
@@ -4,7 +4,7 @@ import ActivityLocation from "@/src/components/ActivityLocation/ActivityLocation
 import ActivityReservationBar from "@/src/components/ActivityReservation/ActivityReservationBar";
 import ActivityTitleSection from "@/src/components/ActivityTitleSection";
 import Loading from "@/src/components/Loading";
-import MyActivityHandler from "@/src/components/MyAtivities/MyActivityHandler";
+import MyActivityHandler from "@/src/components/MyActivities/MyActivityHandler";
 import Review from "@/src/components/Review";
 import Custom404 from "@/src/pages/404";
 import { useGetActivities } from "@/src/queries/useActivities";
